Replace `mach` with `uname -p` in node wrapper script to be more portable
I know I know, this is only meant to be ran on Solaris. But, without
this wrapper-node script(and a few patches), I don't think I ever would have gotten firefox to build on Linux/SPARC. And I believe that this system is extremely useful to anyone else on any nodeless platform.

Thanks for all that you do here. I know maintaining firefox for a mostly
headless and niche platform + for your own OS has got to be frustrating
at times. But, this is extremely helpful, and many thanks from the Tux
Team <3